+++ server/src/main/java/org/allaymc/server/block/component/grass/BlockDryGrassBaseComponentImpl.java
@@ -1,0 +1,23 @@
+package org.allaymc.server.block.component.grass;
+
+import org.allaymc.api.block.BlockBehavior;
+import org.allaymc.api.block.dto.Block;
+import org.allaymc.api.block.type.BlockType;
+import org.allaymc.api.entity.Entity;
+import org.allaymc.api.item.ItemStack;
+
+import java.util.Set;
+
+/**
+ * @author daoge_cmd
+ */
+public class BlockDryGrassBaseComponentImpl extends BlockShortGrassBaseComponentImpl {
+    public BlockDryGrassBaseComponentImpl(BlockType<? extends BlockBehavior> blockType) {
+        super(blockType);
+    }
+
+    @Override
+    public Set<ItemStack> getDrops(Block block, ItemStack usedItem, Entity entity) {
+        return Set.of();
+    }
+}